add record mass upload
